Implement EncodeU128

+++ relayer/crates/cairo-encoding-components/src/impls/encode/mod.rs
@@ -3,4 +3,4 @@ pub mod felt;
 pub mod from_felt;
 pub mod iter;
 pub mod u256;
-pub mod u64;
+pub mod u128;
